Split utils into client and server side
