+++ src/components/dashboard/CalItem.tsx
@@ -1,7 +1,7 @@
 import { useAtom } from "jotai";
 import { calMeetingAtom } from "../atoms/onboardFormAtom";
 import { type FC } from "react";
-import { cn, getMajorClassname } from "@/lib/utils";
+import { cn, getMajorClassname } from "@/lib/clientUtils";
 import { AnimatePresence, motion } from "framer-motion";
 import { Avatar, AvatarFallback, AvatarImage } from "../ui/avatar";
