+++ src/components/ui/alert.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/clientUtils";
 import { X } from "lucide-react";
 
 const alertVariants = cva(
