+++ src/components/ui/label.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as LabelPrimitive from "@radix-ui/react-label";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/clientUtils";
 
 const labelVariants = cva(
   "text-sm font-medium leading-none text-input-text peer-disabled:cursor-not-allowed peer-disabled:opacity-70",
